Cleanup cache structure
