+++ src/main/java/edu/kit/kastel/sdq/lissa/ratlr/postprocessor/TraceLinkIdPostprocessor.java
@@ -42,6 +42,4 @@ public class TraceLinkIdPostprocessor {
         }
         return result;
     }
-
-
 }
